Added require ownership to inventory routes
Added ownership to post to show index, post, patch, delete routes

+++ app/routes/inventory_routes.js
@@ -14,7 +14,7 @@ const customErrors = require('../../lib/custom_errors')
 const handle404 = customErrors.handle404
 // we'll use this function to send 401 when a user tries to modify a resource
 // that's owned by someone else
-// const requireOwnership = customErrors.requireOwnership
+const requireOwnership = customErrors.requireOwnership
 
 // this is middleware that will remove blank fields from `req.body`, e.g.
 // { inventory: { title: '', text: 'foo' } } -> { inventory: { text: 'foo' } }
@@ -22,19 +22,20 @@ const removeBlanks = require('../../lib/remove_blank_fields')
 // passing this as a second argument to `router.<verb>` will make it
 // so that a token MUST be passed for that route to be available
 // it will also set `req.user`
-// const requireToken = passport.authenticate('bearer', { session: false })
+const requireToken = passport.authenticate('bearer', { session: false })
 
 // instantiate a router (mini app that only handles routes)
 const router = express.Router()
 
 // INDEX
 // GET /inventories
-router.get('/inventories', (req, res, next) => {
+router.get('/inventories', requireToken, (req, res, next) => {
   Inventory.find()
     .then(inventories => {
       // `inventories` will be an array of Mongoose documents
       // we want to convert each one to a POJO, so we use `.map` to
       // apply `.toObject` to each one
+      requireOwnership(req, inventories)
       return inventories.map(inventory => inventory.toObject())
     })
     // respond with status 200 and JSON of the inventories
@@ -45,21 +46,24 @@ router.get('/inventories', (req, res, next) => {
 
 // SHOW
 // GET /inventories/5a7db6c74d55bc51bdf39793
-router.get('/inventories/:id', (req, res, next) => {
+router.get('/inventories/:id', requireToken, (req, res, next) => {
   // req.params.id will be set based on the `:id` in the route
   Inventory.findById(req.params.id)
     .then(handle404)
     // if `findById` is succesful, respond with 200 and "inventory" JSON
-    .then(inventory => res.status(200).json({ inventory: inventory.toObject() }))
+    .then(inventory => {
+      requireOwnership(req, inventory)
+      res.status(200).json({ inventory: inventory.toObject() })
+    })
     // if an error occurs, pass it to the handler
     .catch(next)
 })
 
 // CREATE
 // POST /inventories
-router.post('/inventories', (req, res, next) => {
+router.post('/inventories', requireToken, (req, res, next) => {
   // set owner of new inventory to be current user
-  // req.body.inventory.owner = req.user.id
+  req.body.inventory.owner = req.user.id
 
   Inventory.create(req.body.inventory)
     // respond to succesful `create` with status 201 and JSON of new "inventory"
@@ -74,17 +78,17 @@ router.post('/inventories', (req, res, next) => {
 
 // UPDATE
 // PATCH /inventories/5a7db6c74d55bc51bdf39793
-router.patch('/inventories/:id', removeBlanks, (req, res, next) => {
+router.patch('/inventories/:id', requireToken, removeBlanks, (req, res, next) => {
   // if the client attempts to change the `owner` property by including a new
   // owner, prevent that by deleting that key/value pair
-  // delete req.body.inventory.owner
+  delete req.body.inventory.owner
 
   Inventory.findById(req.params.id)
     .then(handle404)
     .then(inventory => {
       // pass the `req` object and the Mongoose record to `requireOwnership`
       // it will throw an error if the current user isn't the owner
-      // requireOwnership(req, inventory)
+      requireOwnership(req, inventory)
 
       // pass the result of Mongoose's `.update` to the next `.then`
       return inventory.updateOne(req.body.inventory)
@@ -97,12 +101,12 @@ router.patch('/inventories/:id', removeBlanks, (req, res, next) => {
 
 // DESTROY
 // DELETE /inventories/5a7db6c74d55bc51bdf39793
-router.delete('/inventories/:id', (req, res, next) => {
+router.delete('/inventories/:id', requireToken, (req, res, next) => {
   Inventory.findById(req.params.id)
     .then(handle404)
     .then(inventory => {
       // throw an error if current user doesn't own `inventory`
-      // requireOwnership(req, inventory)
+      requireOwnership(req, inventory)
       // delete the inventory ONLY IF the above didn't throw
       inventory.deleteOne()
     })
